Exclude dotfiles from org-map-drill-entries scan
If scope = directory, this function gets a list of all org files in
the directory. But when operations like strip data change and do not save a file, they cause a backup dotfile to be created in the directory. org-drill
would see it mid-operation and throw a "Non-existent agenda file" error.